refactor(templates): dedupe parse_template_ref and OCI client construction
Both fetcher.rs and metadata.rs had character-identical copies of
parse_template_ref and the ClientConfig+build_registry_auth block.
Moved both into a new private oci_ref module; tests live there too.
Each call site now gets (client, oci_ref, auth) from build_oci_client.
